backend/cortex-sensoriel/requirements.txt
# Cortex Sensoriel - NEOCORTEX Gateway

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Kafka
aiokafka>=0.10.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
